Final commit 1

+++ src/entities/CardReaderLog.java
@@ -13,14 +13,14 @@ import java.sql.Time;
  */
 public class CardReaderLog extends Log{
 
-    public CardReaderLog(String timecreated, String route, String accountNumber, String checkInLocation, String checkInTime, String checkOutLocation, String checkOutTime, float fareCharged) {
+    public CardReaderLog(String timecreated, String route, String accountNumber, String checkInLocation, String checkInTime, String checkOutLocation, String checkOutTime, Double fareCharged) {
         super(timecreated, route, accountNumber, checkInLocation, checkInTime, checkOutLocation, checkOutTime, fareCharged);
     }
     public CardReaderLog(String timecreated, String route, String accountNumber, String checkInLocation, String checkInTime) {
         super(timecreated, route, accountNumber, checkInLocation, checkInTime);
     }
     
-    public void updateLog(String destination,String time,float fare){
+    public void updateLog(String destination,String time,Double fare){
         this.setCheckOutLocation(destination);
         this.setCheckOutTime(time);
         this.setFareCharged(fare);
